Made some small improvements based on feedback from Will
