+++ garden/src/App.tsx
@@ -64,9 +64,6 @@ function Root() {
 
         {/* Modal Routes */}
         <Route path="modal">
-          <Route element={<PrivateRoutes />}>
-            {/* <Route path="upload" element={<ModalUploadPage />} /> */}
-          </Route>
           <Route path=":id" element={<ModalFunctionPage />} />
         </Route>
